# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rita - Instructor Review & Ranking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /instructors:
    get:
      operationId: listInstructors
      tags: [instructors]
      summary: List/search instructors
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: specialty
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: minScore
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of instructors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorList"
  /instructors/{id}:
    get:
      operationId: getInstructor
      tags: [instructors]
      summary: Get instructor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Instructor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instructor"
        "404":
          description: Not found
  /instructors/{id}/reviews:
    get:
      operationId: getInstructorReviews
      tags: [instructors]
      summary: Get reviews for instructor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
  /instructors/{id}/stats:
    get:
      operationId: getInstructorStats
      tags: [instructors]
      summary: Get instructor score breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorStats"

  # Sessions
...
  /reviews:
    get:
      operationId: listMyReviews
      tags: [reviews]
      summary: List my submitted reviews
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Rankings
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Instructor:
      type: object
      required: [id, name, specialty, verified, avgScore, reviewCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        specialty:
          type: string
        photoUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        verified:
          type: boolean
        avgScore:
          type: number
        avgTechnique:
          type: number
        avgCommunication:
          type: number
        avgPatience:
          type: number
        avgAdaptability:
          type: number
        avgExpertise:
          type: number
        reviewCount:
          type: integer
        publicRank:
          type: ["integer", "null"]
        createdAt:
          type: string
    InstructorList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Instructor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    InstructorStats:
      type: object
      required: [instructorId, avgTechnique, avgCommunication, avgPatience, avgAdaptability, avgExpertise, avgScore, reviewCount, scoreDistribution]
      properties:
        instructorId:
          type: integer
        avgTechnique:
          type: number
        avgCommunication:
          type: number
        avgPatience:
          type: number
        avgAdaptability:
          type: number
        avgExpertise:
          type: number
        avgScore:
          type: number
        reviewCount:
          type: integer
        scoreDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ScoreBucket"
    ScoreBucket:
      type: object
      required: [score, count]
      properties:
        score:
          type: integer
        count:
          type: integer

    # Sessions
...
    Review:
      type: object
      required: [id, userId, instructorId, technique, communication, patience, adaptability, expertise, overallScore, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        userAvatarUrl:
          type: ["string", "null"]
        instructorId:
          type: integer
        instructorName:
          type: ["string", "null"]
        sessionId:
          type: ["integer", "null"]
        technique:
          type: number
        communication:
          type: number
        patience:
          type: number
        adaptability:
          type: number
        expertise:
          type: number
        overallScore:
          type: number
        comment:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
    ReviewList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ReviewInput:
      type: object
      required: [instructorId, technique, communication, patience, adaptability, expertise]
      properties:
        instructorId:
          type: integer
        sessionId:
          type: integer
        technique:
          type: number
          minimum: 1
          maximum: 5
        communication:
          type: number
          minimum: 1
          maximum: 5
        patience:
          type: number
          minimum: 1
          maximum: 5
        adaptability:
          type: number
          minimum: 1
          maximum: 5
        expertise:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string

    # Rankings
...
    Notification:
      type: object
      required: [id, userId, type, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string

    # Dashboard
...