# lib/api-zod/src/generated/types/instructorList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rita - Instructor Review & Ranking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Instructor } from './instructor';

export interface InstructorList {
  items: Instructor[];
  total: number;
  page: number;
  limit: number;
}